More comprehensive build fixes for recent changes

+++ source/numeric/SVInt.cpp
@@ -631,7 +631,7 @@ SVInt SVInt::replicate(const SVInt& times) const {
 }
 
 size_t SVInt::hash() const {
-    return detail::hashing::hash(getRawData(), getNumWords() * WORD_SIZE);
+    return (size_t)detail::hashing::hash(getRawData(), getNumWords() * WORD_SIZE);
 }
 
 std::ostream& operator<<(std::ostream& os, const SVInt& rhs) {
